+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Add `Tooltip` helper component.
+
 ### Changed
 
 - `Pointer` component now support also `onPointerEnter` and `onPointerLeave`